--- src/main/java/chess/controller/web/dao/GameDao.java
@@ -1,4 +1,0 @@
-package chess.controller.web.dao;
-
-public class GameDao {
-}
