+++ src/main/java/chess/dao/GameDao.java
@@ -1,0 +1,4 @@
+package chess.dao;
+
+public class GameDao {
+}
